+++ plugin/src/main/java/com/iluwatar/plugin/ChatApplication.java
@@ -20,19 +20,48 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.amitinside.plugin_pattern.service.impl;
+package com.iluwatar.plugin;
 
-import com.amitinside.plugin_pattern.service.IChatService;
+import com.google.inject.Inject;
+import com.google.inject.name.Named;
+import com.iluwatar.plugin.annotations.Google;
+import com.iluwatar.plugin.service.IChatService;
 
 /**
- * Facebook Chat Service implementation
+ * Chat Application consuming services from different chat service providers
  */
-public final class FacebookChatService implements IChatService {
+public final class ChatApplication {
 
-  /** {@inheritDoc}} */
-  @Override
-  public String chat() {
-    return "Using Facebook Chat";
+  /**
+   * Actual Chat Service
+   */
+  @Inject
+  @Google
+  private IChatService chatService;
+
+  /**
+   * The user id of the chat initiator
+   */
+  @Inject
+  @Named("user")
+  private String userId;
+
+  /**
+   * Starts the chat messaging
+   *
+   * @return the string representation of the actual service call
+   */
+  public String startChat() {
+    return this.chatService.chat();
+  }
+
+  /**
+   * Returns the user id of the messaging initiator
+   *
+   * @return the user id
+   */
+  public String whoIsChatting() {
+    return this.userId;
   }
 
 }
