+++ plugin/src/main/java/com/iluwatar/plugin/Client.java
@@ -16,11 +16,11 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.amitinside.plugin_pattern;
+package com.iluwatar.plugin;
 
-import com.amitinside.plugin_pattern.modules.ProductionModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
+import com.iluwatar.plugin.config.modules.ProductionModule;
 
 /**
  * Client Application
